MODTLR-30 Fix creation of DCB transactions (#31)
